+++ gaffney-advisors/app/components/AboutHero.tsx
@@ -73,7 +73,7 @@ export default function AboutHero() {
           left: 0,
           right: 0,
           bottom: 0,
-          backgroundImage: 'url("/stlouis-skyline.jpg")',
+            // backgroundImage: 'url("/gwmlogo.png")',
           backgroundSize: 'cover',
           backgroundPosition: 'center',
           filter: 'brightness(0.9)',