+++ gaffney-advisors/app/theme-provider.tsx
@@ -128,7 +128,7 @@ const themeOptions: ThemeOptions = {
       fontWeight: 600,
     },
   },
-  spacing: [0, 4, 8, 12, 16, 24, 32, 48, 64, 96, 128, 192, 256],
+  spacing: (factor: number) => `${factor * 4}px`,
   breakpoints: {
     values: {
       xs: 0,